refactor(#451): migrate useCronValidation hook to TypeScript
